requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.1.0
cycler==0.11.0
fonttools==4.40.0
frozendict==2.3.8
frozenlist==1.3.3
html5lib==1.1
idna==3.4
iniconfig==2.0.0
joblib==1.2.0
kiwisolver==1.4.4
lxml==4.9.2
matplotlib==3.7.1
multidict==6.0.4
multitasking==0.0.11
nltk==3.8.1
numpy==1.25.0
pandas==2.0.2
Pillow==9.5.0
pluggy==1.2.0
praw==7.7.0
prawcore==2.3.0
prettytable==3.8.0
pyparsing==3.1.0
pytest==7.4.0
python-dotenv==1.0.0
pytz==2023.3
regex==2023.6.3
requests==2.31.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
soupsieve==2.4.1
threadpoolctl==3.1.0
tqdm==4.65.0
tzdata==2023.3
update-checker==0.18.0
urllib3==2.0.3
webencodings==0.5.1
websocket-client==1.6.1
yarl==1.9.2
yfinance==0.2.22
